+++ app/(menu)/player/coach/page.tsx
@@ -1,5 +1,5 @@
 import { getCoachList } from "@/api/player/apis";
-import { CoachList } from "@/components/player/coach-list";
+import { CoachList } from "@/components/player/coach/coach-list";
 export const dynamic = "force-dynamic";
 export default async function Page() {
   const coachList = await getCoachList();
