+++ app/(menu)/player/pitcher/page.tsx
@@ -1,5 +1,5 @@
 import { getPitcherList } from "@/api/player/apis";
-import PitcherList from "@/components/player/pitcher-list";
+import PitcherList from "@/components/player/pitcher/pitcher-list";
 export const dynamic = "force-dynamic";
 export default async function page() {
   const pitcherList = await getPitcherList();
