+++ components/player/coach/coach-detail.tsx
